feat: sync exchange
