+++ src/frontend/src/lib/schema/canister.schema.ts
@@ -1,0 +1,3 @@
+import { PrincipalTextSchema } from '$lib/schema/principal.schema';
+
+export const CanisterIdTextSchema = PrincipalTextSchema;
