+++ src/frontend/src/lib/services/auth.services.ts
@@ -3,7 +3,12 @@ import { resetSubnets } from '$lib/services/subnets.services';
 import { authStore, type AuthSignInParams } from '$lib/stores/auth.store';
 import { busy } from '$lib/stores/busy.store';
 import { i18n } from '$lib/stores/i18n.store';
-import {cyclesIdbStore, exchangeIdbStore, monitoringIdbStore, statusesIdbStore} from '$lib/stores/idb.store';
+import {
+	cyclesIdbStore,
+	exchangeIdbStore,
+	monitoringIdbStore,
+	statusesIdbStore
+} from '$lib/stores/idb.store';
 import { toasts } from '$lib/stores/toasts.store';
 import type { ToastLevel, ToastMsg } from '$lib/types/toast';
 import { replaceHistory } from '$lib/utils/route.utils';
